refactor(ui): extract sync-config presentational pieces (APIC-P3-11)
Split DirDiagram / EndpointSelect / ReadonlyValue out of SyncConfigScreen into
sync-components.tsx so the page stays under the 500-line admin monolith guard
(ADR-0021) and reads as a thin composition.

Refs #1789
